Print a message when we start the Zuul console
The link for the job console is made available before Ansible starts
printing things. If the console is empty, users will be greeted with
an "END OF STREAM" message and will need to keep refreshing until there
is actual content.

Writing a minimal message to "start" the console will prevent that.

Change-Id: I4200c91b468b1c2fcccee23dda19962b30335ef5