Don't shrink windows on reconfiguration
The window size supplied by the user on a pipeline is really
the starting window size.  It can grow without bound, and so
there is no reason to shrink a window on reconfiguration.
For that matter, it's not necessary to increase it either,
unless the user has raised the floor higher than the current
value.

Therefore, use the existing window value on static (ie, dependent)
change queues during reconfiguration, or the new floor if it is
higher.

Dynamic queues will get new values from the pipeline (whose value
is also conditionally updated in the same manner).

In review, we discovered that some folks may be using the expotential
window type to create a static window.  To make sure this doesn't
break that case, add a test for it.  Since that can be used to
intentionally reduce the window size on reconfiguration, this new
test mirrors the test for window shrinkage added in the previous
change which must be altered now that we don't auto-shrink on
reconfiguration.

Change-Id: Iaba25788ebe51ced919dc896aa20aa90675d7773