Fix complex branch matchers in project configs
A recent change to always apply an implied branch matcher to in-tree
project configs on projects with branches had an optimization to
avoid adding unecessary jobs which was too simple.  It assumed that
if the raw text of the matcher on the job did not match the name
of the branch, then there was no need to add the job.  That only
works for branch matchers that are simple branch names.  One that
is a regex may or may not match the implied branch.

To correct this, when building the job list, run the implied branch
name against the branch matcher regex (if there is one) for the job
to determine whether it will not match.

Change-Id: Ibbe097801a45928bc9942991d868a78f5f441887