Web: Drop some columns from build list
The newrev, node name, and end time are not as useful as the other
columns, and can be found by following links to the change or logs.
In the future, we can redesign the display so that the user can select
extra details for a change, but for now, improve the ability to
scan the results quickly by removing these less-used columns.

Change-Id: I17179178b5486101c9134f8684c6e69f35e1ea7c