Handle sigterm in nodaemon mode
When running zuul within a container it normally runs in nodaemon mode
as pid 1. Currently in this mode zuul just ignores SIGTERM which is
used normally to stop containers. Thus when running within OpenShift
it waits for a timeout until it gets killed forcefully.

Fix this by handling SIGINT and SIGTERM equally.

Change-Id: I24bd8c953e734fdb9545714126d77cbcdc161bbd