Reprime the installation map in getGitUrl if needed
The getGitUrl is also used by the mergers and executors. In order to
work with github apps they need to know the mapping of projects to
installation ids. However as they don't receive webhook events they
currently only create this mapping on startup. Thus we need to refresh
the installation map if the project is not found.

Change-Id: I4e0f4473e0402ea0d619b83e56eb2a3f4fbe6698