Tenant config can be read from an external script
This patch adds a new option for the scheduler "tenant_config_script".
When this option is set and target an executable script then the script
is executed and its stdout is taken as the tenant config. This new
option is excluse with "tenant_config". When set, "tenant_config"
is ignored.

The goal is to be able to fetch tenant configuration from
an external source whatever the type of the source as
the logic is handled in a script.

Change-Id: Ic8234dfcdf1f07f824542c4a2ebf99fe8ff19fd5