Add change information to Build Completed log message
Sometimes when tracking issues it's useful to be able to correlate
builds with the executor on which they ran. Currently there are two
lines:

  2018-01-22 06:25:15,198 DEBUG zuul.Pipeline.openstack.check: Build
    <Build b54719f8a45c4f738baece9323a7a87d of openstack-tox-py35 on
    <Worker ze04.openstack.org>> completed
  2018-01-22 06:25:15,198 DEBUG zuul.Pipeline.openstack.check: Item
    <QueueItem 0x7f301cb019b0 for <Change 0x7f3032d117b8 535656,6> in check>
    status is now:

which have pretty much all the information you'd need, but doesn't allow
grepping for change and jobname and getting the info about where it ran
and how to find it.

By adding Item into the line, we get:

  2018-01-22 06:25:15,198 DEBUG zuul.Pipeline.openstack.check: Build
    <Build b54719f8a45c4f738baece9323a7a87d of openstack-tox-py35 on
    <Worker ze04.openstack.org>> of <Change 0x7f3032d117b8 535656,6> completed

Which allows simple grepping for a change + job and finding the worker and
build.

Change-Id: I3239d0fcb4614fe295edc55b34bb2419127d1174