Update docs to use sphinx-build
Update the tox.ini helper env 'docs' to use sphinx-build directly like
the job does.

Use the build-sphinx-docs job rather than the openstack version as we do
not need constraints in zuul.

This does not move doc requirements to doc/requirements.txt because the
Zuul docs document the tests too, so the doc build needs all the things
in test-requirements as well.

Change-Id: Id81c05725d4dbe12dccc0272c396c92ff2f475ec