Use override-checkout to select jobs
If a job is defined in a project with "master" and "stable" branches,
and a user wants to inherit from that job in a project with only a
"devel" branch, Zuul will not run the job because there are no matching
variants.  To correct this, have Zuul use the value of override-checkout
to mask the branch of the item under test so that it appears to match the
'override-checkout' branch.

Change-Id: I4619a6f9fe9c3203a7285e5eb4f74c52b650c737