Fix self fulfilling empty node requests
Emtpy node requests are intended to be fulfilled directly in zuul
without bugging zk. However we didn't check for this condition in
acceptNodes. This causes us to 'recreate' the missing empty node
request on zk, let nodepool fulfill it and accept this afterwards. Fix
this by short-circuiting acceptNodes if we have an empty and fulfilled
node request.

Change-Id: Ib953d366258e5724c85ae2a563e319422f4ada40