Don't use run_lock in executor's merger
The run_lock is used to ensure that the executor has a consistent
list of jobs running.  Specifically, that we don't start a new
job right as we're shutting down and stopping jobs.

When we broke out the executor-merger into its own gearman worker,
we copied the run_lock into that as well.  This means that as long
as a merger job is running, an executor job can't start.

The merger doesn't actually need the run_lock, so remove it.

Change-Id: Ibb96c7e5b6d02962170fbe0c611c1f2665002c47