Sync when doing disk accountant testing
Depending on your filesystem and kernel you may not actually sync things
to disk quickly enough in the tests to cause the disk accountant tests
to pass reliably. Update the tests to sync to disk explicitly so that
the tests are reliable.

This was noticed as being a problem on btrfs /tmp on suse tumbleweed.

Change-Id: I38fce053ee878751540b4c9d4b05863dc5588de8