Use nested tempfile fixture for cleanups
There are a couple places where we appear to leak tempfile and tempdirs
because we are creating dirs/files in the tests then not deleting them
when we are done. Address this by using the nested tempfile fixture
which will nest all tempfiles and tempdirs in a tempdir that the fixture
cleans up when the test is completed.

Change-Id: I2818ed48823c544bb1be6b4e1e58f78109197bc1