Remove .json suffix from web routes
For the existing simple cases, like builds, jobs and status, having the json
suffix is a perfectly reasonable thing. However, in the next patch it starts
to get weird. When we add support for specific changes or specific jobs, we
we grow URLs like:

  /openstack/status.json
  /openstack/status/change/537010,2.json

Those read weird, because change/537010,2 is much more like an argument or
specialiation of status. The thing that reads weird is the status call having
 .json but the change-specific call just being status/, not
the trailing .json on the change url.

Removing the json suffix gets us:

  /openstack/status
  /openstack/status/change/537010,2

which feels better as the status portion of the url remains consistent.

This is done first in the stack so that as we add tests for new
endpoints we can get them right the first time rather than having a big
rename patch at the end (which is what this started as)

Change-Id: I4baf33fdacaf46943fbd192743551bb27bd618de