Use a status code to detect unknown vs. missing tenant
Re-use http status codes for this, since it's just arbitrary
machine-reable info for now anyway and that way we can just splat the
payload into the HTTPNotFound when it's missing.
Use 204 for the tenant not being ready yet, since it's a valid request
there just isn't any data right now.

Change-Id: Ife3de557d48a12c85671643877106479e77d7fd6