Import Zuul modules at top of files
We had to make sure to not import Paramiko before daemonization.  The
zuul-server would hang when establishing a Gerrit ssh connection due to
Random.Crypto() failling to acquire random number from /dev/urandom. It
would block on read() and never process.

When the Server command line invokes the daemonization, python-daemon
closes all file descriptors. Including /dev/urandom. Then the daemonized
establishes the SSH connection and fails to get random number because
Random.Crypto() locks on read() on a closed file description.

Paramiko issue is https://github.com/paramiko/paramiko/issues/59 and the
fix is to use os.random:
https://github.com/paramiko/paramiko/commit/6f211115f49edcea7d23b764d7cf3a84ff12f5f0

That has been released with Paramiko 1.11.6 and we now require 2.0+.

Move Zuul imports at top of files and drop the comments hinting at the
Paramiko bug.

Change-Id: I5fe956df74815761e3eac2bd25f6fd7f167fc854