Actually honor the static_path config value
The puppet in puppet-zuul puts the static files somewhere that isn't
inside the python path and sets static_path to point to it.

Unfortunately, the refactor to split some handler base classes into the
handlers file wound up with StaticHandler just using the STATIC_DIR
default.

Instead, let's use the value set on the zuul_web instance, since that
will either be the config value or STATIC_DIR as a fallback.

Change-Id: Ie512ceb8c9348749c9ad40ee4057950bbb2766ea