Use bootstrap 3.1.1 instead of 3.3.7
The original status page before the webpack work was using bootstrap
3.1.1. In the webpack-ification, we bumped that to 3.3.7 and it seemed
fine. However, the x in the search bar stopped working with 3.3.7.

Pin back to 3.3.1 until we can figure out what went wrong there.

Also remove an extra matcher in webpack.common.js that with 3.3.1 causes
the glyphicons fonts to not be loaded properly.

Change-Id: Ibd41b62166fed4104f9569934d2eca36b338b711