Upgrade to webpack 4
The webpack community just released webpack v4. Upgrade to it.

Amongst the relevant upstream changes, the uglify plugin and
commonschunk plugins are no longer needed as their behavior is now just
builtin. The webpack.LoaderOptionsPlugin is no longer needed, it existed
to ease transition from webpack 1 to webpack 2. minify is now an option
to optimization.

Change-Id: I0359f69861211231e4875abf9817316d9e8878a1