Rename javascript package to @zuul-ci/dashboard
The python service is "zuul-web", but the javascript portion is
currently also being built as "zuul-web". While the javascript code is
intended to be served by zuul-web, not having its own name has made it
awkward to be clear about which codebase is being discussed.

npm has also added namespacing to published modules, which allows
grouping modules and is indicated by the @ prefix.

Update the config to call the javascript code '@zuul-ci/dashboard'. This
will allow us to add an npm publication job, although we need to set up
some accounts for that first.

Change-Id: I6e31ae23f74f887448f0a48147d91c156025e925