Add zuul_json tests
This adds a test of zuul_json to the 'remote' test suite.  We could
probably test this aspect of zuul_json at least in the normal unit
test suite, however, the streaming callback has some behavior which
can only be tested remotely, and it might be nice for all of the
callback tests to behave in the same way.

Change-Id: If9aa0d65e60b011b0c69da582c9bdc2d2bb9a8a6